chore: move wallets and personas to independent data source
