+++ src/social-network/defaults/ui.ts
@@ -4,14 +4,13 @@ import type { SocialNetworkUIDataSources } from '../ui'
 import { ValueRef } from '@holoflows/kit'
 import { ProfileIdentifier } from '../../database/type'
 import { cloneDeep } from 'lodash-es'
-import type { Profile, Group, Persona } from '../../database'
-import { MaskbookDarkTheme, MaskbookLightTheme } from '../../utils/theme'
-import { ProfileArrayComparer, GroupArrayComparer, PersonaArrayComparer } from '../../utils/comparer'
+import type { Profile } from '../../database'
+import { MaskbookLightTheme } from '../../utils/theme'
+import { ProfileArrayComparer, GroupArrayComparer } from '../../utils/comparer'
 
 const defaultDataSources: Required<SocialNetworkUIDataSources> = cloneDeep({
     friendsRef: new ValueRef([], ProfileArrayComparer),
     myIdentitiesRef: new ValueRef([], ProfileArrayComparer),
-    myPersonasRef: new ValueRef([], PersonaArrayComparer),
     groupsRef: new ValueRef([], GroupArrayComparer),
     currentIdentity: new ValueRef<Profile | null>(null),
     lastRecognizedIdentity: new ValueRef({ identifier: ProfileIdentifier.unknown }),
